feat: Rotate camera angle.
rotate camera angle with key.code R and it will move to the right. Use key.code Q it will move it to the left.

We can change this but for now this is how to look around the map and get a different perspective.

The camera rotation transition is a too quick. Next build we can make it smoother.

Any other ideas to handle it add some comments. We could have the camera follow the direction of kento, but I think this would be more of a mario 64 feel to our game.